0.0.4

+++ nhk_easy/api.py
@@ -1,6 +1,14 @@
 import requests
 import time
 from lxml import etree as le
+from copy import deepcopy
+import os
+import logging
+
+logger = logging.getLogger(__file__)
+logger.setLevel(logging.INFO)
+streamhandler = logging.StreamHandler()
+logger.addHandler(streamhandler)
 
 
 class Api(object):
@@ -15,17 +23,24 @@ class Api(object):
         r = requests.get(url)
         return r.json()
 
-    def download_top_news(self):
+    def download_top_news(self, text=True, m3u8=False, mp3=True):
         for news in self.top_news:
             [id_, dttm, title] = [
                 news[a] for a in ("news_id", "news_prearranged_time", "title")
             ]
+            logger.info("Downloading: " + title)
             fn = dttm.split(" ")[0] + "-" + title
             article = Article(id_)
-            with open(fn + ".m3u8", "w") as f:
-                f.write(article.m3u8)
-            with open(fn + ".txt", "w") as f:
-                f.write(article.text)
+            if m3u8 and not os.path.exists(fn + ".m3u8"):
+                m3u8 = article.get_m3u8()
+                with open(fn + ".m3u8", "w") as f:
+                    f.write(m3u8)
+            if text and not os.path.exists(fn + ".txt"):
+                text = article.get_text()
+                with open(fn + ".txt", "w") as f:
+                    f.write(text)
+            if mp3 and not os.path.exists(fn + ".mp3"):
+                article.download_mp3(fn + ".mp3")
             time.sleep(0.5)
 
 
@@ -35,12 +50,12 @@ class Article(object):
 
     def __init__(self, id_):
         self.id_ = id_
-        self.m3u8 = self.get_m3u8()
-        self.text = self.get_text()
+        self.m3u8_url = (
+            f"https://nhks-vh.akamaihd.net/i/news/easy/{self.id_}.mp4/master.m3u8"
+        )
 
     def get_m3u8(self):
-        url = f"https://nhks-vh.akamaihd.net/i/news/easy/{self.id_}.mp4/master.m3u8"
-        r = requests.get(url)
+        r = requests.get(self.m3u8_url)
         return r.text
 
     def get_raw_html(self):
@@ -55,9 +70,5 @@ class Article(object):
         # //div[@id_='js-article-body']
         return "".join(html.xpath("//div[@id='js-article-body']//text()")).strip()
 
-
-# remove_tags = tree.xpath('//div[last()]')
-
-# if remove_tags:
-#     remove_tag = remove_tags[0]
-#     remove_tag.getparent().remove(remove_tag)
+    def download_mp3(self, fn):
+        os.system(f"ffmpeg -i {self.m3u8_url} -acodec mp3 -ab 257k '{fn}'")
